requirements.txt
Django==4.0.4
djangorestframework==3.14.0
django-filter==22.1
django-cors-headers==4.3.1
requests==2.31.0
psycopg2==2.9.6
redis==4.6.0
celery=5.3.6
python-dotenv==1.0.0